Decoupled post service into service and utils.
